// src/config/models/market-price.ts
export interface MarketPrice{
    "symbol":  string; // "BNBBTC"
    "priceChange":  string; // "-94.99999800"
    "priceChangePercent":  string; // "-95.960"
    "weightedAvgPrice":  string; // "0.29628482"
    "prevClosePrice":  string; // "0.10002000"
    "lastPrice":  string; // "4.00000200"
    "lastQty":  string; // "200.00000000"
    "bidPrice":  string; // "4.00000000"
    "bidQty":  string; // "100.00000000"
    "askPrice":  string; // "4.00000200"
    "askQty":  string; // "100.00000000"
    "openPrice":  string; // "99.00000000"
    "highPrice":  string; // "100.00000000"
    "lowPrice":  string; // "0.10000000"
    "volume":  string; // "8913.30000000"
    "quoteVolume":  string; // "15.30000000"
    "openTime":  number; // 1499783499040
    "closeTime":  number; // 1499869899040
    "firstId":  number; // 28385   // First tradeId
    "lastId":  number; // 28460    // Last tradeId
    "count":  number; // 76         // Trade count
}